+++ src/esm/utils/algorithms/Metaballs.d.ts
@@ -10,7 +10,6 @@
  * @date    2024-03-01
  */
 import { Circle } from "../../Circle";
-import { XYCoords } from "../../interfaces";
 interface MetaballsOptions {
     metaRadiusAddon: number;
 }
@@ -36,7 +35,6 @@ export declare class Metaballs {
          * @returns
          */
         detectHoleGroup: (circles: Array<Circle>, nonVisitedSet: Set<number>, index: number) => Array<number>;
-        anyCircleContainsPoint: (circles: Array<Circle>, point: XYCoords, ignoreCircleIndex: number) => boolean;
     };
 }
 export {};