+++ src/ts/CircleSector.ts
@@ -5,13 +5,16 @@
  * @modified 2021-02-26 Fixed an error in the svg-arc-calculation (case angle<90deg and anti-clockwise).
  * @modified 2024-01-30 Added a missing type in the `describeSVGArc` function.
  * @modified 2024-03-01 Added the `getStartPoint` and `getEndPoint` methods.
+ * @modified 2024-03-08 Added the `containsAngle` method.
  * @version  1.2.0
  **/
 
 import { Circle } from "./Circle";
+import { Line } from "./Line";
 import { UIDGenerator } from "./UIDGenerator";
 import { Vertex } from "./Vertex";
 import { SVGPathParams, SVGSerializable, UID, XYCoords } from "./interfaces";
+import { CircularIntervalSet } from "./utils/datastructures/CircularIntervalSet";
 
 /**
  * @classdesc A simple circle sector: circle, start- and end-angle.
@@ -84,6 +87,24 @@ export class CircleSector implements SVGSerializable {
   }
 
   /**
+   * Checks wether the given angle (must be inside 0 and PI*2) is contained inside this sector.
+   *
+   * @param {number} angle - The numeric angle to check.
+   * @method containsAngle
+   * @instance
+   * @memberof CircleSector
+   * @return {boolean} True if (and only if) this sector contains the given angle.
+   */
+  containsAngle(angle: number) {
+    if (this.startAngle <= this.endAngle) {
+      return angle >= this.startAngle && angle < this.endAngle;
+    } else {
+      // startAngle > endAngle
+      return angle >= this.startAngle || angle < this.endAngle;
+    }
+  }
+
+  /**
    * Get the sectors starting point (on the underlying circle, located at the start angle).
    *
    * @method getStartPoint
@@ -106,6 +127,26 @@ export class CircleSector implements SVGSerializable {
   getEndPoint(): Vertex {
     return this.circle.vertAt(this.endAngle);
   }
+
+  // circleSectorIntersection(sector: CircleSector): CircleSector | null {
+  //   const radicalLine: Line | null = this.circle.circleIntersection(sector.circle);
+  //   if (!radicalLine) {
+  //     // The circles to not intersect at all.
+  //     return null;
+  //   }
+  //   // Circles intersect. Check if sector intervals intersect, too.
+  //   const thisIntersectionAngleA = this.circle.center.angle(radicalLine.a);
+  //   const thisIntersectionAngleB = this.circle.center.angle(radicalLine.b);
+  //   // Is intersection inside this sector?
+  //   const thisIntervals: CircularIntervalSet = new CircularIntervalSet(0, Math.PI * 2.0);
+  //   thisIntervals.intersect(thisIntersectionAngleA, thisIntersectionAngleB);
+
+  //   const intersectionSector = new CircleSector(
+  //     new Circle(this.circle.center.clone(), this.circle.radius),
+  //     this.startAngle,
+  //     this.endAngle
+  //   );
+  // }
 
   /**
    * This function should invalidate any installed listeners and invalidate this object.